+++ src/careamics/dataset_ng/batch_sampler/fifo_batch_sampler.py
@@ -1,3 +1,5 @@
+from collections.abc import Sequence
+from pathlib import Path
 from typing import TYPE_CHECKING, Optional
 
 import numpy as np
@@ -8,9 +10,7 @@ if TYPE_CHECKING:
     from ..patch_extractor.image_stack import ManagedLazyImageStack
 
 
-# TODO: decouple from CAREamicsDataset by having n_files as argument
-#   Also have to give patch indices as argument
-class GroupedBatchSampler(BatchSampler):
+class FifoBatchSampler(BatchSampler):
 
     def __init__(
         self,
@@ -61,9 +61,7 @@ class GroupedBatchSampler(BatchSampler):
 
             # yield batches, if not equal to batch size, it is set to remaining indices
             for i in range(0, len(patch_indices), self.batch_size):
-                batch_indices = (
-                    (patch_indices[i : i + self.batch_size]).astype(int).tolist()
-                )
+                batch_indices = list(patch_indices[i : i + self.batch_size])
                 if len(batch_indices) == self.batch_size:
                     yield batch_indices
                 else:
@@ -75,3 +73,54 @@ class GroupedBatchSampler(BatchSampler):
                 yield remaining_indices
 
     # TODO: need to make sure all remaining files are closed at the end
+
+
+class FifoImageStackManager:
+
+    def __init__(self, max_files_loaded: int = 1):
+        self.image_stacks: dict[Path, ManagedLazyImageStack] = {}
+        self.loaded_image_stacks: list[ManagedLazyImageStack] = []
+        self.max_files_loaded = max_files_loaded
+
+    @property
+    def currently_loaded(self):
+        return len(self.loaded_image_stacks)
+
+    def register_image_stack(self, image_stack: "ManagedLazyImageStack"):
+        self.image_stacks[image_stack.path] = image_stack
+        image_stack.set_callbacks(
+            on_load=self.notify_load,
+            on_close=self.notify_close,
+        )
+
+    def register_image_stacks(self, image_stacks: Sequence["ManagedLazyImageStack"]):
+        for image_stack in image_stacks:
+            self.register_image_stack(image_stack)
+
+    def free(self):
+        if self.currently_loaded >= self.max_files_loaded:
+            image_stack_to_close = self.loaded_image_stacks[0]  # FIFO
+            image_stack_to_close.deallocate()
+
+    def notify_load(self, image_stack: "ManagedLazyImageStack"):
+        if image_stack.path not in self.image_stacks:
+            raise ValueError(
+                f"Image stack with path {image_stack.path} has not been registered."
+            )
+        # TODO: Raise error if already loaded?
+        self.free()
+        self.loaded_image_stacks.append(image_stack)
+
+    def notify_close(self, image_stack: "ManagedLazyImageStack"):
+        if image_stack.path not in self.image_stacks:
+            raise ValueError(
+                f"Image stack with path {image_stack.path} has not been registered."
+            )
+        loaded_paths = [img_stack.path for img_stack in self.loaded_image_stacks]
+        file_id = loaded_paths.index(image_stack.path)  # TODO: this will be file path
+        self.loaded_image_stacks.pop(file_id)
+
+    def close_all(self):
+        loaded_image_stacks = self.loaded_image_stacks.copy()
+        for image_stack in loaded_image_stacks:
+            image_stack.deallocate()
